// lib/generated/prisma/models/Vendor.ts
export type VendorModel = runtime.Types.Result.DefaultSelection<Prisma.$VendorPayload>
// ...
export type $VendorPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Vendor"
  objects: {
    Karyawan: Prisma.$KaryawanPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    slugVendor: string
    namaVendor: string
    alamat: string
    noTelp: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["vendor"]>
  composites: {}
}